rename core.modules.json_new -> core.modules.json and core.modules.xml_clean -> core.modules.xml
I think initially it was like this because I was running it like  `bleanser/core/modules/xml.py prune ...`

This would implicitly add src/bleanser/core to PYTHONPATH, and if you have any modules conflicting with builtin modules, it would shadow them (like xml or json), and result in all sorts of trouble (e.g. segfaults).

Running as `python3 -m bleanser.core.modules.xml prune ...` is much cleaner anyway and results in less trouble in other aspects as well, so let's just do that and rename the core modules for the ease of discovery
