+++ src/bleanser/core/modules/xml.py
@@ -1,8 +1,4 @@
 #!/usr/bin/env python3
-"""
-Ugh, wtf?? If I name it simply 'xml', I get all sorts of weird behaviours... presumably because it conflicts with some system modules..
-"""
-
 from lxml import etree
 
 from contextlib import contextmanager
