+++ src/bleanser/modules/xml_clean.py
@@ -1,4 +1,4 @@
-from bleanser.core.modules.xml_clean import *  # noqa: F401, F403
+from bleanser.core.modules.xml import *  # noqa: F401, F403
 
 import warnings
 warnings.warn("Module 'bleanser.modules.xml_clean' is deprecated. Use 'bleanser.core.modules.xml_clean' instead.", DeprecationWarning)
